Cleanup formatter module
And fixes a bug with keywords for examples.
