+++ .generator/src/generator/templates/example.j2
@@ -5,9 +5,9 @@
     {# Include 'from os import environ' used in given statements. #}
     {%- set given = context._imports["os"].add("environ") -%}
 {%- endif -%}
-{%- set parameters, imports = format_parameters_python(context.api_request.kwargs, spec=operation_spec, replace_values=context._replace_values, version=version) %}
+{%- set parameters, imports = format_parameters(context.api_request.kwargs, spec=operation_spec, replace_values=context._replace_values, version=version) %}
 {%- if context.body %}
-{%- set body, imports = format_data_with_schema_python(context.body.value, context.api_request.schema.spec, replace_values=context._replace_values, version=version, imports=imports) %}
+{%- set body, imports = format_data_with_schema(context.body.value, context.api_request.schema.spec, replace_values=context._replace_values, version=version, imports=imports) %}
 {%- endif %}
 
 {%- for package, values in context._imports.items() %}
