+++ examples/v1/synthetics/CreateSyntheticsAPITest_1487281163.py
@@ -32,7 +32,7 @@ body = SyntheticsAPITest(
         assertions=[
             SyntheticsAssertionTarget(
                 operator=SyntheticsAssertionOperator("is"),
-                property="{{ PROPERTY }}",
+                _property="{{ PROPERTY }}",
                 target="text/html",
                 type=SyntheticsAssertionType("header"),
             ),
