+++ examples/v1/synthetics/UpdateAPITest.py
@@ -36,7 +36,7 @@ body = SyntheticsAPITest(
         assertions=[
             SyntheticsAssertionTarget(
                 operator=SyntheticsAssertionOperator("is"),
-                property="{{ PROPERTY }}",
+                _property="{{ PROPERTY }}",
                 target="text/html",
                 type=SyntheticsAssertionType("header"),
             ),
